+++ autoadmin/__init__.py
@@ -1,14 +1,11 @@
-from django.core.exceptions import ImproperlyConfigured
-
-try:
-    from .handlers import autoadmin_create, autoadmin_account_passwd_change  # NOQA
-except ImproperlyConfigured:
-    # Ignore this exception during setup.py
-    pass
+from __future__ import unicode_literals
 
 __author__ = 'Roberto Rosario'
-__build__ = 0x000500
+__build__ = 0x010000
 __copyright__ = 'Copyright 2014 Roberto Rosario'
 __license__ = 'MIT'
 __title__ = 'django-autoadmin'
-__version__ = '0.5.0'
+__version__ = '1.0.0'
+
+default_app_config = 'autoadmin.apps.AutoAdminAppConfig'
+