+++ autoadmin/models.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-from django.contrib.auth.models import User
+from django.conf import settings
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
@@ -8,7 +8,7 @@ from solo.models import SingletonModel
 
 
 class AutoAdminSingleton(SingletonModel):
-    account = models.ForeignKey(User, blank=True, null=True,
+    account = models.ForeignKey(settings.AUTH_USER_MODEL, blank=True, null=True,
                                 verbose_name=_('Account'))
     password = models.CharField(blank=True, max_length=128, null=True,
                                 verbose_name=_('Password'))