+++ autoadmin/settings.py
@@ -3,8 +3,6 @@ from __future__ import unicode_literals
 from django.conf import settings
 from django.contrib.auth.models import User
 
-EMAIL = getattr(settings, 'AUTOADMIN_EMAIL', 'autoadmin@autoadmin.com')
-PASSWORD = getattr(settings, 'AUTOADMIN_PASSWORD',
-                   User.objects.make_random_password())
+EMAIL = getattr(settings, 'AUTOADMIN_EMAIL', 'autoadmin@example.com')
+PASSWORD = getattr(settings, 'AUTOADMIN_PASSWORD', None)
 USERNAME = getattr(settings, 'AUTOADMIN_USERNAME', 'admin')
-ENABLE = getattr(settings, 'AUTOADMIN_ENABLE', True)