+++ HISTORY.rst
@@ -1,3 +1,19 @@
+1.0.0 (2015-06-15)
+==================
+
+- Update included partial template
+- Django admin interface enabled
+- Drop support for Django < 1.6
+- Adds support Django >= 1.7
+- Autoadmin user is no longer created after database migration
+- Add new createautoadmin management command
+- Uses Django's own 
+- Adds Django native database migrations
+- Uses AppConfig to about running code during import
+- Adds supports for custom User models
+- ENABLE setting removed
+- Default email changed to autoadmin@example.com
+
 0.5.0 (2014-08-25)
 ==================